requirements.txt
SQLAlchemy==1.4.37	
anyio==3.6.1	
click==8.1.3	
colorama==0.4.5
cx-Oracle==8.3.0	
et-xmlfile==1.1.0	
greenlet==1.1.2
h11==0.13.0	
idna==3.3	
openpyxl==3.0.10	
pip~=22.2.2
psycopg2==2.9.3	
pydantic==1.9.1	
setuptools==62.6.0	
sniffio==1.2.0	
starlette==0.19.1	
typing-extensions==4.2.0	

python-dotenv~=0.21.0
attrs~=22.1.0
fastapi~=0.78.0
requests~=2.28.1
uvicorn~=0.18.1